+++ metrics/src/main/scala/org/broadinstitute/dsde/workbench/metrics/package.scala
@@ -18,4 +18,6 @@ package object metrics extends LazyLogging {
       .build(host, port)
     reporter.start(period.toMillis, period.toMillis, TimeUnit.MILLISECONDS)
   }
+
+  implicit def metricUnwrapper[M](metric: Metric[M]): M = metric.metric
 }